+++ scripts/macros/2024/classFeatures/warlock/greatOldOne/eldritchHex.js
@@ -1,0 +1,5 @@
+export let eldritchHex = {
+    name: 'Eldritch Hex',
+    version: '1.5.21',
+    rules: 'modern'
+};